disk: separate most tests from package
Separate most disk tests into internal and external ones.

- disk_test.go: was internal (package disk) and is now external (package
  disk_test)
- Function `entityPath()` and method
  `PartitionTable.findDirectoryEntityPath()` are now exported for tests
  in export_test.go
- Test partition tables moved to partition_table_internal_test.go and
  exported.
- Unique string tests moved to new file: disk_internal_test.go
