+++ pkg/disk/export_test.go
@@ -1,3 +1,10 @@
 package disk
 
-var PayloadEntityMap = payloadEntityMap
+var (
+	PayloadEntityMap = payloadEntityMap
+	EntityPath       = entityPath
+)
+
+func FindDirectoryEntityPath(pt *PartitionTable, path string) []Entity {
+	return pt.findDirectoryEntityPath(path)
+}
